Verifying that task CRUD work

+++ app/src/main/java/com/job/softclick_mobile/adapters/EmployeeSelectItemAdapter.java
@@ -79,7 +79,7 @@ public class EmployeeSelectItemAdapter extends BaseAdapter {
         }
     }
 
-    public int getPosition(int id) {
+    public int getPosition(Long id) {
         AtomicInteger position = new AtomicInteger();
         employees.forEach(e -> {
             int i = employees.indexOf(e);
